Favoritar e desfavoritar cifras pela tela de cifra (#3352)
* adicionar use case para favoritar e desfavoritar versions
request da api do desfavoritar
inserir comportamento na tela de version

* testes e alguns pequenos ajustes

* corrigir erro ao deletar da lista de favorito pela tela de lista

* correção do review

* remove current snackbar

* fix teste não conclusivo

+++ lib/domain/songbook/use_cases/delete_versions.dart
@@ -1,4 +1,5 @@
 import 'package:cifraclub/domain/shared/request_error.dart';
+import 'package:cifraclub/domain/songbook/models/list_type.dart';
 import 'package:cifraclub/domain/songbook/repository/songbook_repository.dart';
 import 'package:cifraclub/domain/songbook/repository/user_songbook_repository.dart';
 import 'package:cifraclub/domain/version/models/version.dart';
@@ -12,9 +13,17 @@ class DeleteVersions {
   final UserVersionRepository _userVersionRepository;
   final UserSongbookRepository _userSongbookRepository;
 
-  DeleteVersions(this._songbookRepository, this._userVersionRepository, this._userSongbookRepository);
+  const DeleteVersions(
+    this._songbookRepository,
+    this._userVersionRepository,
+    this._userSongbookRepository,
+  );
 
   Future<Result<void, RequestError>> call({required int songbookId, required List<Version> versions}) async {
+    if (ListType.getListTypeById(songbookId) != ListType.user) {
+      throw Exception("Don't support delete version from this list type");
+    }
+
     final songsRemoteIds = versions.map((e) => e.remoteDatabaseId ?? -1).toList();
 
     return _songbookRepository.deleteVersions(songbookId: songbookId, versionsId: songsRemoteIds).then(
